Reorg EitherTraverse
