+++ kategory/src/main/kotlin/kategory/data/EitherT.kt
@@ -68,13 +68,13 @@ data class EitherT<F, A, B>(val MF: Monad<F>, val value: HK<F, Either<A, B>>) : 
             OptionT(MF, MF.map(value, { it.toOption() }))
 
     fun <C> foldL(b: C, f: (C, B) -> C, FF: Foldable<F>): C =
-            FF.compose(EitherTraverse<A>()).foldLC(value, b, f)
+            FF.compose(Either.foldable<A>()).foldLC(value, b, f)
 
     fun <C> foldR(lb: Eval<C>, f: (B, Eval<C>) -> Eval<C>, FF: Foldable<F>): Eval<C> =
-            FF.compose(EitherTraverse<A>()).foldRC(value, lb, f)
+            FF.compose(Either.foldable<A>()).foldRC(value, lb, f)
 
     fun <G, C> traverse(f: (B) -> HK<G, C>, GA: Applicative<G>, FF: Traverse<F>, MF: Monad<F>): HK<G, HK<EitherTF<F, A>, C>> {
-        val fa = ComposedTraverse(FF, EitherTraverse<A>(), Either.monad<A>()).traverseC(value, f, GA)
+        val fa = ComposedTraverse(FF, Either.traverse<A>(), Either.monad<A>()).traverseC(value, f, GA)
         return GA.map(fa, { EitherT(MF, MF.map(it.lower(), { it.ev() })) })
     }
 }
