+++ kategory/src/main/kotlin/kategory/instances/EitherInstances.kt
@@ -4,7 +4,9 @@ interface EitherInstances<L> :
         Functor<EitherF<L>>,
         Applicative<EitherF<L>>,
         Monad<EitherF<L>>,
-        MonadError<EitherF<L>, L>{
+        MonadError<EitherF<L>, L>,
+        Foldable<EitherF<L>>,
+        Traverse<EitherF<L>> {
 
     override fun <A> pure(a: A): Either<L, A> = Either.Right(a)
 
@@ -34,5 +36,29 @@ interface EitherInstances<L> :
             is Either.Right -> fea
         }
     }
+
+    override fun <G, A, B> traverse(fa: HK<EitherF<L>, A>, f: (A) -> HK<G, B>, GA: Applicative<G>): HK<G, HK<EitherF<L>, B>> =
+            fa.ev().let { either ->
+                when (either) {
+                    is Either.Right -> GA.map(f(either.b), { Either.Right(it) })
+                    is Either.Left -> GA.pure(either)
+                }
+            }
+
+    override fun <A, B> foldL(fa: HK<EitherF<L>, A>, b: B, f: (B, A) -> B): B =
+            fa.ev().let { either ->
+                when (either) {
+                    is Either.Right -> f(b, either.b)
+                    is Either.Left -> b
+                }
+            }
+
+    override fun <A, B> foldR(fa: HK<EitherF<L>, A>, lb: Eval<B>, f: (A, Eval<B>) -> Eval<B>): Eval<B> =
+            fa.ev().let { either ->
+                when (either) {
+                    is Either.Right -> f(either.b, lb)
+                    is Either.Left -> lb
+                }
+            }
 
 }
