+++ kategory/src/test/kotlin/kategory/data/EitherTTest.kt
@@ -221,7 +221,7 @@ class EitherTTest : UnitSpec() {
             val traverse = eitherT.traverse(f, Option, Id, Id).ev()
             val result = traverse.map { it.ev().value.value() }
 
-            val expected = EitherTraverse<String>().traverse(either, f, Option)
+            val expected = Either.traverse<String>().traverse(either, f, Option)
             result shouldBe expected
         }
